// platform/HarmonyOS/entry/.preview/default/cache/default/default@PreviewArkTS/esmodule/debug/entry/src/main/ets/model/CicadaTypes.ts
export interface MultiPlayerSelection {
    url: string;
    name: string;
}